Moved MatchListener to net.ninebolt.onevsone.match
