+++ src/main/java/net/ninebolt/onevsone/OneVsOne.java
@@ -21,9 +21,9 @@ import net.ninebolt.onevsone.command.RootCommand;
 import net.ninebolt.onevsone.event.CacheUniqueIdListener;
 import net.ninebolt.onevsone.event.CreateStatsListener;
 import net.ninebolt.onevsone.event.MatchEndEvent;
-import net.ninebolt.onevsone.event.MatchListener;
 import net.ninebolt.onevsone.match.Match;
 import net.ninebolt.onevsone.match.MatchEndCause;
+import net.ninebolt.onevsone.match.MatchListener;
 import net.ninebolt.onevsone.match.MatchManager;
 import net.ninebolt.onevsone.match.MatchSelector;
 import net.ninebolt.onevsone.stats.Stats;
